rename `on-rails.js` to `cypress-on-rails.js`
