+++ lib/generators/cypress_on_rails/templates/spec/cypress/support/cypress-on-rails.js
