+++ lib/generators/cypress_on_rails/templates/spec/cypress/support/index.js
@@ -15,7 +15,7 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands'
-import './on-rails'
+import './cypress-on-rails'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
